ai-agents/06b-build-remote-agents-with-a2a/python/requirements.txt
python-dotenv
httpx
azure-identity
azure-ai-projects
openai
aiohttp
uvicorn
starlette
sse-starlette
fastapi
